Migrated utility funcs to TS
